add AverageScore functionality to the restaurant show view

+++ BestRestaurant/Models/Rating.cs
@@ -4,7 +4,7 @@ namespace BestRestaurant.Models
   {
     public int RatingId { get; set; }
     public int RestaurantId { get; set; }
-    public int Score { get; set; }
+    public decimal Score { get; set; }
     public string Review { get; set; }
     public virtual Restaurant Restaurant { get; set; }
   }
